fix(offline-messaging): move OfflineUser into User.swift and fix MessageEvent init
- OfflineUser was in a separate file not tracked by the Xcode project, causing
  archive build failures; merged into User.swift which is already in the target
- Fixed MessageEvent initializer call: eventID/userID/icon → id/senderUserID/senderIcon
