+++ Wired 3/Models/User.swift
@@ -160,6 +160,20 @@ struct TransferDisplaySnapshot {
 
 @Observable
 @MainActor
+final class OfflineUser: Identifiable {
+    let login: String
+    var nick: String
+
+    var id: String { login }
+
+    init(login: String, nick: String? = nil) {
+        self.login = login
+        self.nick = nick ?? login
+    }
+}
+
+@Observable
+@MainActor
 final class User: Identifiable {
     let id: UInt32
     var nick: String
